Refactor css with nesting syntax
