+++ example/src/Components/Tile/index.css
@@ -2,16 +2,14 @@
     width: 49vw;
     height: calc(30vw - 2px);
     transition: width 1s ease;
-}
-
-.tile img {
-    width: 100%;
-    height: 100%;
-    object-fit: cover;
-}
-
-.tile img:empty {
-    background-color: #efefef;
-    width: 49vw;
-    height: calc(30vw - 2px);
+    img {
+        width: 100%;
+        height: 100%;
+        object-fit: cover;
+        &:empty {
+            background-color: #efefef;
+            width: 49vw;
+            height: calc(30vw - 2px);
+        }
+    }
 }
