+++ example/src/Screens/Home/index.css
@@ -1,31 +1,27 @@
 .home {
     background-color: #efefef;
     flex: 1;
-}
-.home .list {
-    background-color: white;
-    display: flex;
-    flex-direction: column;
-}
-.home .list button {
-    padding: 0;
-}
-.home .motion-element {
-    position: absolute;
-    bottom: 25px;
-    left: 50%;
-}
-.home .list {
-    list-style: none;
-    padding: 0;
-}
-.home .list li {
-    display: flex;
-}
-.home .list a {
-    flex: 1;
-    text-decoration: none;
-}
-.home .list button {
-    width: 100%;
+    .list {
+        list-style: none;
+        padding: 0;
+        background-color: white;
+        display: flex;
+        flex-direction: column;
+        li {
+            display: flex;
+        }
+        a {
+            flex: 1;
+            text-decoration: none;
+        }
+        button {
+            width: 100%;
+            padding: 0;
+        }
+    }
+    .motion-element {
+        position: absolute;
+        bottom: 25px;
+        left: 50%;
+    }
 }
