+++ example/src/Screens/Overlays/Home/index.css
@@ -1,0 +1,46 @@
+.overlays .modal-example {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+.overlays .go-back {
+    position: absolute;
+    z-index: 1;
+    transform: translate(15px, calc(env(safe-area-inset-top) + 12px));
+}
+
+.player {
+    width: calc(100vw - 30px);
+    height: calc(46px + env(safe-area-inset-bottom));
+    display: flex;
+    padding: 15px;
+    padding-bottom: 5vh;
+    justify-content: space-between;
+    background-color: #222222;
+    color: white;
+    position: absolute;
+    bottom: 0;
+    left: 0;
+    .cover-art {
+        width: 46px;
+        height: 46px;
+        overflow: hidden;
+        img {
+            object-fit: cover;
+            width: 100%;
+            height: 100%;
+            border-radius: 5px;
+        }
+    }
+    .info {
+        display: flex;
+        flex: 1;
+        gap: 10px;
+        align-items: center;
+    }
+    .play-controls {
+        display: flex;
+        gap: 8px;
+        align-items: center;
+    }
+}
