+++ example/src/Screens/Overlays/Home/index.tsx
@@ -2,11 +2,11 @@ import { Button, IconButton } from '@mui/material';
 import React, { useRef } from 'react';
 import { Anchor, SharedElement } from '@react-motion-router/core';
 import { Stack } from '@react-motion-router/stack';
-import King from "../../assets/king.webp";
+import King from "../../../assets/king.webp";
 import SkipNextIcon from '@mui/icons-material/SkipNext';
 import PlayArrowIcon from '@mui/icons-material/PlayArrow';
 import ArrowBackIosIcon from '@mui/icons-material/ArrowBackIos';
-import '.././index.css';
+import './index.css';
 
 interface OverlaysProps extends Stack.ScreenComponentProps { }
