+++ example/src/Screens/Overlays/index.css
@@ -1,34 +1,5 @@
 @import "./modal.css";
 
-.overlays.suspense > * {
-  animation: overlays-fade-in .5s ease forwards;
-}
-.overlays .navbar {
-    background-color: transparent;
-    color: transparent;
-}
-.overlays dialog.animation-provider {
-    border: none;
-    background-color: transparent;
-    padding: 0px;
-    max-width: unset;
-}
-.overlays {
-    flex: 1;
-    display: flex;
-    flex-direction: column;
-    background-color: white;
-}
-.overlays .modal-example {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-.overlays .go-back {
-    position: absolute;
-    z-index: 1;
-    transform: translate(15px, calc(env(safe-area-inset-top) + 12px));
-}
 @keyframes overlays-fade-in {
     from {
         opacity: 0;
@@ -40,38 +11,22 @@
 #overlays {
     overflow: hidden;
 }
-.player {
-    width: calc(100vw - 30px);
-    height: calc(46px + env(safe-area-inset-bottom));
-    display: flex;
-    padding: 15px;
-    padding-bottom: 5vh;
-    justify-content: space-between;
-    background-color: #222222;
-    color: white;
-    position: absolute;
-    bottom: 0;
-    left: 0;
-}
-.player .cover-art {
-    width: 46px;
-    height: 46px;
-    overflow: hidden;
-}
-.player .cover-art img {
-    object-fit: cover;
-    width: 100%;
-    height: 100%;
-    border-radius: 5px;
-}
-.player .info {
-    display: flex;
+.overlays {
     flex: 1;
-    gap: 10px;
-    align-items: center;
-}
-.player .play-controls {
     display: flex;
-    gap: 8px;
-    align-items: center;
+    flex-direction: column;
+    background-color: white;
+    &.suspense > * {
+        animation: overlays-fade-in .5s ease forwards;
+    }
+    dialog.animation-provider {
+        border: none;
+        background-color: transparent;
+        padding: 0px;
+        max-width: unset;
+    }
+    .navbar {
+        background-color: transparent;
+        color: transparent;
+    }
 }
